+++ markwon-core/src/main/java/io/noties/markwon/MarkwonVisitor.java
@@ -23,6 +23,19 @@ public interface MarkwonVisitor extends Visitor {
         void visit(@NonNull MarkwonVisitor visitor, @NonNull N n);
     }
 
+    /**
+     * Primary purpose is to control the spacing applied before/after certain blocks, which
+     * visitors are created elsewhere
+     *
+     * @since $nap;
+     */
+    interface BlockHandler {
+
+        void blockStart(@NonNull MarkwonVisitor visitor, @NonNull Node node);
+
+        void blockEnd(@NonNull MarkwonVisitor visitor, @NonNull Node node);
+    }
+
     interface Builder {
 
         /**
@@ -32,6 +45,16 @@ public interface MarkwonVisitor extends Visitor {
          */
         @NonNull
         <N extends Node> Builder on(@NonNull Class<N> node, @Nullable NodeVisitor<? super N> nodeVisitor);
+
+        /**
+         * @param blockHandler to handle block start/end
+         * @see BlockHandler
+         * @see BlockHandlerDef
+         * @since $nap;
+         */
+        @SuppressWarnings("UnusedReturnValue")
+        @NonNull
+        Builder blockHandler(@NonNull BlockHandler blockHandler);
 
         @NonNull
         MarkwonVisitor build(@NonNull MarkwonConfiguration configuration, @NonNull RenderProps renderProps);
@@ -133,4 +156,14 @@ public interface MarkwonVisitor extends Visitor {
      */
     @SuppressWarnings("unused")
     <N extends Node> void setSpansForNodeOptional(@NonNull Class<N> node, int start);
+
+    /**
+     * @since $nap;
+     */
+    void blockStart(@NonNull Node node);
+
+    /**
+     * @since $nap;
+     */
+    void blockEnd(@NonNull Node node);
 }